Changed the abbreviated state name (WA) to the full name (Western Australia) to be consistent with other holiday providers.
